make shared review page part 2
