+++ publisher/src/components/ReviewMetrics/ReviewMetrics.tsx
@@ -15,104 +15,233 @@
 // along with this program.  If not, see <https://www.gnu.org/licenses/>.
 // =============================================================================
 
-import { observer } from "mobx-react-lite";
-import React, { useEffect, useState } from "react";
-import { useLocation, useNavigate, useParams } from "react-router-dom";
+/* eslint-disable camelcase */
 
-import { UploadedMetric } from "../DataUpload/types";
-import { REPORTS_LOWERCASE } from "../Global/constants";
-import { ReviewModal } from "../Reports/ReviewModal";
+import { DatapointsTableView } from "@justice-counts/common/components/DataViz/DatapointsTableView";
+import { formatDateShortMonthYear } from "@justice-counts/common/components/DataViz/utils";
+import { useIsFooterVisible } from "@justice-counts/common/hooks";
+import { RawDatapoint, ReportOverview } from "@justice-counts/common/types";
+import React, { useState } from "react";
+import { useNavigate, useParams } from "react-router-dom";
+
+import { printReportTitle } from "../../utils";
+import checkIcon from "../assets/check-icon.svg";
+import logoImg from "../assets/jc-logo-vector-new.svg";
+import errorIcon from "../assets/status-error-icon.png";
 import {
-  ReviewHeaderActionButton,
-  ReviewMetricOverwrites,
-  SharedReviewMetrics,
-} from "./SharedReviewMetrics";
+  Button,
+  ButtonTypes,
+  ReviewMetricsButtonsContainer,
+  ReviewMetricsHeader,
+} from "../DataUpload";
+import {
+  REPORT_CAPITALIZED,
+  REPORTS_CAPITALIZED,
+  REPORTS_LOWERCASE,
+} from "../Global/constants";
+import { Logo, LogoContainer } from "../Header";
+import { EmptyIcon } from "../Reports/PublishConfirmation.styles";
+import {
+  Heading,
+  HeadingGradient,
+  MetricsPanel,
+  MetricStatusIcon,
+  ReviewMetricsWrapper,
+  SectionContainer,
+  SectionExpandStatusSign,
+  Summary,
+  SummarySection,
+  SummarySectionLine,
+  SummarySectionsContainer,
+  SummarySectionTitle,
+} from "./ReviewMetrics.styles";
 
-const ReviewMetrics: React.FC = observer(() => {
+export type ReviewHeaderActionButton = {
+  name: string;
+  type: ButtonTypes;
+  onClick: () => void;
+  disabled?: boolean;
+};
+
+export type ReviewMetric = {
+  datapoints: RawDatapoint[];
+  display_name: string;
+  key: string;
+  metricHasError?: boolean;
+  metricHasValidInput?: boolean;
+};
+
+export type ReviewMetricOverwrites = {
+  key: number;
+  metricName: string;
+  dimensionName: string;
+  startDate: string;
+};
+
+type ReviewMetricsProps = {
+  title: string;
+  description: string | React.ReactNode;
+  buttons: ReviewHeaderActionButton[];
+  metrics: ReviewMetric[];
+  metricOverwrites?: ReviewMetricOverwrites[];
+  records?: ReportOverview[];
+};
+
+export const ReviewMetrics: React.FC<ReviewMetricsProps> = ({
+  title,
+  description,
+  buttons,
+  metrics,
+  metricOverwrites,
+  records,
+}) => {
   const { agencyId } = useParams();
-  const { state } = useLocation();
-  const { uploadedMetrics, fileName } = state as {
-    uploadedMetrics: UploadedMetric[] | null;
-    fileName: string;
-  };
   const navigate = useNavigate();
-  const [isSuccessModalOpen, setIsSuccessModalOpen] = useState(false);
+  const isFooterVisible = useIsFooterVisible();
+  const [isMetricsSectionExpanded, setIsMetricsSectionExpanded] =
+    useState(true);
+  const [isOverwritesSectionExpanded, setIsOverwritesSectionExpanded] =
+    useState(true);
+  const [isRecordsSectionExpanded, setIsRecordsSectionExpanded] = useState(() =>
+    records ? records.length <= 10 : false
+  );
 
-  useEffect(() => {
-    if (!uploadedMetrics) {
-      // no metrics in passed in navigation state, redirect to home page
-      navigate(`/agency/${agencyId}/${REPORTS_LOWERCASE}`, { replace: true });
-    }
-  });
-
-  if (!uploadedMetrics) {
-    return null;
-  }
-
-  const filteredMetrics = uploadedMetrics
-    .map((metric) => ({
-      ...metric,
-      datapoints: metric.datapoints.filter((dp) => dp.value !== null),
-    }))
-    .filter((metric) => metric.datapoints.length > 0);
-
-  const overwrites: ReviewMetricOverwrites[] = [];
-  filteredMetrics.forEach((metric) => {
-    metric.datapoints.forEach((dp) => {
-      if (dp.old_value !== null) {
-        const overwriteData: ReviewMetricOverwrites = {
-          key: dp.id,
-          metricName: dp.metric_display_name || "",
-          dimensionName: dp.dimension_display_name || "",
-          startDate: dp.start_date,
-        };
-        overwrites.push(overwriteData);
-      }
-    });
-  });
-
-  const title =
-    filteredMetrics.length > 0 ? "Review Upload" : "No Metrics to Review";
-
-  const description =
-    filteredMetrics.length > 0
-      ? "Here’s a breakdown of data from the file you uploaded. You can publish these changes now, or save as a draft for later."
-      : "Uploaded file contains no metrics to review.";
-
-  const buttons: ReviewHeaderActionButton[] =
-    filteredMetrics.length > 0
-      ? [
-          {
-            name: "Exit Without Publishing",
-            type: "border",
-            onClick: () => navigate(`/agency/${agencyId}/${REPORTS_LOWERCASE}`),
-          },
-          {
-            name: "Publish",
-            type: "green",
-            onClick: () => setIsSuccessModalOpen(true),
-          },
-        ]
-      : [
-          {
-            name: "Close",
-            type: "red",
-            onClick: () => navigate(-1),
-          },
-        ];
+  const renderSection = (metric: ReviewMetric) => {
+    return (
+      <SectionContainer key={metric.key}>
+        <DatapointsTableView
+          datapoints={metric.datapoints}
+          metricName={metric.display_name}
+        />
+      </SectionContainer>
+    );
+  };
 
   return (
-    <>
-      {isSuccessModalOpen && <ReviewModal fileName={fileName} />}
-      <SharedReviewMetrics
-        title={title}
-        description={description}
-        buttons={buttons}
-        metrics={filteredMetrics}
-        metricOverwrites={overwrites}
-      />
-    </>
+    <ReviewMetricsWrapper>
+      <ReviewMetricsHeader transparent={false}>
+        <LogoContainer
+          onClick={() => navigate(`/agency/${agencyId}/${REPORTS_LOWERCASE}`)}
+        >
+          <Logo src={logoImg} alt="" />
+        </LogoContainer>
+        <ReviewMetricsButtonsContainer>
+          {buttons.map(({ name, type, onClick, disabled }) => (
+            <Button
+              key={name}
+              type={type}
+              onClick={onClick}
+              disabled={disabled}
+            >
+              {name}
+            </Button>
+          ))}
+        </ReviewMetricsButtonsContainer>
+        {
+          // TODO(#24): Add Publish button to publish multiple reports at once (todo from upload metrics)
+        }
+      </ReviewMetricsHeader>
+      <Summary isFooterVisible={isFooterVisible}>
+        <Heading>
+          {title}
+          <span>{description}</span>
+        </Heading>
+        <SummarySectionsContainer>
+          <HeadingGradient />
+          {metrics.length > 0 && (
+            <SummarySection>
+              <SummarySectionTitle
+                color="blue"
+                onClick={() =>
+                  setIsMetricsSectionExpanded(!isMetricsSectionExpanded)
+                }
+              >
+                <span>{metrics.length}</span> Metric
+                {metrics.length > 1 ? "s" : ""}
+                <SectionExpandStatusSign>
+                  {isMetricsSectionExpanded ? "-" : "+"}
+                </SectionExpandStatusSign>
+              </SummarySectionTitle>
+              {isMetricsSectionExpanded &&
+                metrics.map(
+                  ({
+                    key,
+                    display_name,
+                    metricHasError,
+                    metricHasValidInput,
+                  }) => (
+                    <SummarySectionLine key={key}>
+                      {!metricHasError && metricHasValidInput && (
+                        <MetricStatusIcon src={checkIcon} alt="" />
+                      )}
+                      {metricHasError && (
+                        <MetricStatusIcon src={errorIcon} alt="" />
+                      )}
+                      {!metricHasError && !metricHasValidInput && <EmptyIcon />}
+                      {display_name}
+                    </SummarySectionLine>
+                  )
+                )}
+            </SummarySection>
+          )}
+          {metricOverwrites && metricOverwrites.length > 0 && (
+            <SummarySection>
+              <SummarySectionTitle
+                color="orange"
+                onClick={() =>
+                  setIsOverwritesSectionExpanded(!isOverwritesSectionExpanded)
+                }
+              >
+                <span>{metricOverwrites.length}</span> Overwrite
+                {metricOverwrites.length > 1 ? "s" : ""}
+                <SectionExpandStatusSign>
+                  {isOverwritesSectionExpanded ? "-" : "+"}
+                </SectionExpandStatusSign>
+              </SummarySectionTitle>
+              {isOverwritesSectionExpanded &&
+                metricOverwrites.map(
+                  ({ key, metricName, dimensionName, startDate }) => (
+                    <SummarySectionLine key={key}>
+                      {metricName}: {dimensionName}
+                      <span>({formatDateShortMonthYear(startDate)})</span>
+                    </SummarySectionLine>
+                  )
+                )}
+            </SummarySection>
+          )}
+          {records && records.length > 0 && (
+            <SummarySection>
+              <SummarySectionTitle
+                color="grey"
+                onClick={() =>
+                  setIsRecordsSectionExpanded(!isRecordsSectionExpanded)
+                }
+              >
+                <span>{records.length}</span>{" "}
+                {records.length > 1 ? REPORTS_CAPITALIZED : REPORT_CAPITALIZED}
+                <SectionExpandStatusSign>
+                  {isRecordsSectionExpanded ? "-" : "+"}
+                </SectionExpandStatusSign>
+              </SummarySectionTitle>
+              {isRecordsSectionExpanded &&
+                records.map((record) => (
+                  <SummarySectionLine key={record.id}>
+                    {printReportTitle(
+                      record.month,
+                      record.year,
+                      record.frequency
+                    )}
+                  </SummarySectionLine>
+                ))}
+            </SummarySection>
+          )}
+        </SummarySectionsContainer>
+      </Summary>
+      <MetricsPanel>
+        {metrics.map((metric) => {
+          return renderSection(metric);
+        })}
+      </MetricsPanel>
+    </ReviewMetricsWrapper>
   );
-});
-
-export default ReviewMetrics;
+};
